Fix rebase errors
